requirements.txt
vanna
python-dotenv
qdrant-client
openai
fastembed
sql_metadata
streamlit
tiktoken
sqlglot